Task/skpath replace svg

+++ Scaffold.Core/CalcQuantities/CalcStress.cs
@@ -110,7 +110,7 @@ public sealed class CalcStress : CalcQuantity<Stress>
 
     public static CalcStress operator /(CalcStress x, int y)
     {
-        return new CalcStress(x.Value / (double)y, (StressUnit)x.Quantity.Unit, x.DisplayName, x.Symbol);
+        return new CalcStress(x.Value / y, (StressUnit)x.Quantity.Unit, x.DisplayName, x.Symbol);
     }
 
     public static CalcStress operator /(CalcStress x, CalcDouble y)
